Add files via upload
Description:
			
Firstly, we will declare two arrays of size 12 to store months of years and declare suitable variables.
Then we will take input from user of last year and current year rainfall statistics on same indexes of both arrays. The input will be taken in the form of pairs.
Then we will show the table of last year and current year rainfall statistics.
Then we will calculate total rainfall in last year and current year by adding the values of 12 months of both years separately.
Then we will calculate average rainfall of both years by dividing their total to 12 respectively.
Then we will find which months have highest rainfall in a year by using for loop and if statement in it and comparing one month with all months of the last year and current year respectively. Then showing month using for loop for both years.
Then we will find which month have no rainfall in both years. We will check all the months if it has zero (0) in it we will say that there is no rainfall in that month. We will show the month name by using for loop.
Then we will predict in which months of next year rainfall will occur. We will add the rainfall statistics of same month of both last year and current year respectively.
•	If sum of both months is greater than 5 we will take average by dividing it with 2 and cout that value as the prediction that rainfall will occur in that month.
•	If sum of both months is less than 5 we will cout zero (0) as a prediction that rainfall will not occur in that month.
In the end we will show a table of months of last year, current year, and next year and show their values.